backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Google Cloud
google-cloud-bigquery==3.14.1
google-auth==2.27.0

# Scheduler
apscheduler==3.10.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
